:sparkles: implement "overview mode" (#27)
